+++ basics.js
@@ -1,4 +1,4 @@
-Game.spawns["Spawn1"].spawnCreep([WORK, WORK, CARRY, MOVE], "harvester-" + Game.time, {
+Game.spawns["Spawn2"].spawnCreep([WORK, WORK, CARRY, MOVE], "harvester-" + Game.time, {
   memory: { role: "harvester" }
 });